+++ lamb/__init__.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 __author__ = 'KoNEW'
 
-__version__ = '2.4.5'
+__version__ = '2.4.6'
 
 
 import logging